Restore linux_386 support.
	modified:   Makefile
	modified:   arpa/inet/const_linux_386.go
	modified:   crt_linux_386.go
	modified:   crt_linux_amd64.go
	modified:   ctype/const_linux_386.go
	modified:   dirent/const_linux_386.go
	copied:     dirent_linux_amd64.go -> dirent_linux.go
	modified:   dirent_linux_amd64.go
	modified:   fcntl_linux_386.go
	modified:   float/const_linux_386.go
	modified:   limits/const_linux_386.go
	modified:   math/const_linux_386.go
	modified:   netdb/const_linux_386.go
	modified:   pthread/const_linux_386.go
	modified:   pthread_linux.go
	modified:   signal/const_linux_386.go
	modified:   stdlib/const_linux_386.go
	modified:   sys/mman/const_linux_386.go
	modified:   sys/param/const_linux_386.go
	modified:   sys/resource/const_linux_386.go
	modified:   sys/socket/const_linux_386.go
	modified:   sys/time/const_linux_386.go
	modified:   sys/wait/const_linux_386.go
	modified:   sys_stat_linux_386.go
	modified:   unistd/const_linux_386.go
	modified:   zlib/const_linux_386.go

+++ pthread_linux.go
@@ -950,7 +950,7 @@ func Xpthread_cond_timedwait(tls TLS, cond, mutex, abstime uintptr) (r int32) {
 	ts := *(*Stimespec)(unsafe.Pointer(abstime))
 	if ts.Xtv_nsec < 0 || ts.Xtv_nsec > 1e9 {
 	}
-	deadline := time.Unix(ts.Xtv_sec, ts.Xtv_nsec)
+	deadline := time.Unix(int64(ts.Xtv_sec), int64(ts.Xtv_nsec))
 	if r = Xpthread_cond_wait(tls, cond, mutex); r != 0 {
 		return r
 	}
